switching plugin manager to nvim.lazy, initial commit to dev (#67)
* WIP: starting work on lazy load plugin


* feat: WIP commiting a batch of lazy configurations.

* WIP: another lazy batch


* feat: wip, another lazy commit


* feat: wip, another lazy commit


* WIP: chore more lazy changes


* wip: lazy work on telescope and which-key


* feat: lazy, 0.1 on lazy


---------
