+++ lua/mengano/core/init.lua
@@ -1,0 +1,4 @@
+require("mengano.core.nvim-options")
+require("mengano.core.winbar")
+require("mengano.core.mappings")
+require("mengano.core.autocommands")
