+++ lua/mengano/core/nvim-options.lua
@@ -53,7 +53,7 @@ local global_options = {
     mouse = "n",
     ruler = true,
     swapfile = false,
-    pastetoggle = "<F2>",
+    -- pastetoggle = "<F2>",
     pumheight = 20,
     redrawtime = 250,
     relativenumber = true,
