+++ lua/mengano/plugins/nvim-autopairs.lua
@@ -1,0 +1,24 @@
+return {
+    'windwp/nvim-autopairs',
+    event = "InsertEnter",
+    opts = {
+        check_ts = true,
+        ts_config = {
+            lua = { "string", "source" },
+            javascript = { "string", "template_string" },
+            java = false,
+        },
+        disable_filetype = { "TelescopePrompt", "spectre_panel" },
+        fast_wrap = {
+            map = "<M-e>",
+            chars = { "{", "[", "(", '"', "'" },
+            pattern = string.gsub([[ [%'%"%)%>%]%)%}%,] ]], "%s+", ""),
+            offset = 0, -- Offset from pattern match
+            end_key = "$",
+            keys = "qwertyuiopzxcvbnmasdfghjkl",
+            check_comma = true,
+            highlight = "PmenuSel",
+            highlight_grey = "LineNr",
+        },
+    } 
+}
