+++ lua/mengano/plugins/nvim-colorizer.lua
@@ -1,0 +1,1 @@
+return { "norcalli/nvim-colorizer.lua" }
