--- lua/nvimtree-config/init.lua
@@ -1,54 +1,0 @@
--- skip if module(s) isn't loaded
-local status_ok, nvimtree = pcall(require, "nvim-tree")
-if not status_ok then
-  return
-end
-
-nvimtree.setup({
-  actions = {
-    open_file = {
-      quit_on_open = true,
-      resize_window = true,
-    },
-  },
-  diagnostics = {
-    enable = true,
-    show_on_dirs = true,
-    icons = {
-      hint = " ",
-      info = " ",
-      warning = " ",
-      error = " ",
-    },
-  },
-  disable_netrw = true,
-  filters = { dotfiles = true },
-  git = {},
-  reload_on_bufenter = true,
-  renderer = {
-    add_trailing = true,
-    indent_markers = {
-      enable = false,
-    },
-    highlight_opened_files = "all",
-    highlight_git = true,
-    icons = {
-      padding = " ",
-      glyphs = {
-        git = {
-          unstaged = "",
-          staged = "S",
-          unmerged = "",
-          renamed = "➜",
-          deleted = "",
-          untracked = "U",
-          ignored = "◌",
-        },
-      },
-    },
-  },
-  sync_root_with_cwd = true,
-  view = {
-    adaptive_size = true,
-  },
-})
