--- lua/user/lsp/init.lua
@@ -1,7 +1,0 @@
-local status_ok, _ = pcall(require, "lspconfig")
-if not status_ok then
-    return
-end
-
-require("user.lsp.language-servers")
-require("user.lsp.diagnostics")
